.history/requirements_20251025004058.txt
# Core Framework
sqlalchemy>=2.0.25

# Environment & Config
python-dotenv==1.0.0

# Data Processing
pandas>=2.0.0
requests>=2.31.0

# Text encoding detection
chardet>=5.0.0

# GUI Framework
# tkinter is included with Python
